Day3
Row 19: Made a script to use jquery using Bower
Row 16: Created a new div with class "resultSection"
Row 20-30: created a own script with a function to hide "resultSection" and then to give a function to the Add-button using inputs and append (jQ-api) input.
On row 28 I connected the function with a click actin (jQ)

When you write a script its done by a $ and note that you finnish with ().  It is also used after } in row 29. Need to figure out why you use {}
In this case its used in row 21, 24,26 and 29