Day4 Tjejerkodar
Index: Row 17 added a class = submit for button
Row 29-30 added variable options and variable submit function
Row 31 hidden the inputSection in the webpage
Row 32 written script to instead show only the result when pressing the Go-button.
Row 36 written in a formula to be able to pick out a winning option from the list. Math.random times the lenght of the options picks random item in the list and Math.floor makes sure to get rid of decimals.
Row 38 script to show the result in text of the winning option.
Row 40 script to submit when clicking button.

In main.css I have added some styling to make the resultSection look nicer.